+++ src/main/java/com/ardmore/quarters/gentlemens/repository/AppointmentRepository.java
@@ -1,12 +1,11 @@
 package com.ardmore.quarters.gentlemens.repository;
 
 import com.ardmore.quarters.gentlemens.entity.Appointment;
+import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-import org.springframework.data.repository.CrudRepository;
-
-public interface AppointmentRepository extends CrudRepository<Appointment, Integer> {
+public interface AppointmentRepository extends CrudRepository<Appointment, Long> {
     List<Appointment> findAllByDateAndTimeAndEmployeeId(String date, String time, int employeeId);
 
     Appointment findByAppointmentId(int id);
